Update rocksDbVersion and update code to not use the workarounds for the fixed bugs.

+++ edge-util/test/Microsoft.Azure.Devices.Edge.Storage.RocksDb.Test/TestRocksDbStoreProvider.cs
@@ -29,7 +29,7 @@ namespace Microsoft.Azure.Devices.Edge.Storage.RocksDb.Test
         public void Dispose()
         {
             this.rocksDbStoreProvider?.Dispose();
-            if (!string.IsNullOrWhiteSpace(this.rocksDbFolder))
+            if (!string.IsNullOrWhiteSpace(this.rocksDbFolder) && Directory.Exists(this.rocksDbFolder))
             {
                 Directory.Delete(this.rocksDbFolder, true);
             }
